pyproject.toml
[project]
name = "obsidian-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "ruff>=0.15.4",
    "watchdog>=4.0.0",
]
